+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-nothing new yet
+### Fixed
+- `YAMLFile` will use an empty dictionary instead of None when loading empty yaml files.
 
 ## [0.1.1] - 2019-07-20
-nothing new yet
+### Added
+- More metadata for the packaging index
 
 ## [0.1.0] - 2019-07-20
 ### Added